// src/common/backend/payments/types.ts
type InvoiceLine = {
    product: string;
    quantity: number;
    price: number;
    cost: number;
};

type CostAmountRange = {
    lowerBound: number;
    upperBound: number;
};

export type InvoiceTemplateLineCostUnlim = {
    costType: 'InvoiceTemplateLineCostUnlim';
};

export type InvoiceTemplateLineCostRange = {
    costType: 'InvoiceTemplateLineCostRange';
    range: CostAmountRange;
    currency: string;
};

export type InvoiceTemplateLineCostFixed = {
    costType: 'InvoiceTemplateLineCostFixed';
    amount: number;
    currency: string;
};

export type InvoiceTemplateSingleLine = {
    templateType: 'InvoiceTemplateSingleLine';
    product: string;
    price: InvoiceTemplateLineCostFixed | InvoiceTemplateLineCostRange | InvoiceTemplateLineCostUnlim;
};

export type InvoiceTemplateMultiLine = {
    templateType: 'InvoiceTemplateMultiLine';
    cart: InvoiceLine[];
    currency: string;
};